Navigator, PathWrapper, UIHandler, Reader
TODO:
 - Read file to start Reader
 - Rename, Delete
 - Shortcuts
