+++ Manga-Reader/Manga-Reader/Manga-Reader.csproj
@@ -49,6 +49,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Container.cs" />
+    <Compile Include="FileContainer.cs" />
+    <Compile Include="FilePage.cs" />
+    <Compile Include="FilePageWrapper.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -67,21 +70,21 @@
     <Compile Include="Form3.Designer.cs">
       <DependentUpon>Form3.cs</DependentUpon>
     </Compile>
-    <Compile Include="IFolder.cs" />
-    <Compile Include="IFile.cs" />
-    <Compile Include="INavigator.cs" />
-    <Compile Include="IPathWrapper.cs" />
-    <Compile Include="IUIHandler.cs" />
     <Compile Include="MyTreeNode.cs" />
     <Compile Include="MyTreeView.cs">
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="Navigator.cs" />
     <Compile Include="Page.cs" />
+    <Compile Include="PageWrapper.cs" />
+    <Compile Include="FilePathWrapper.cs" />
+    <Compile Include="PathWrapper.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Reader.cs" />
+    <Compile Include="Readere.cs" />
     <Compile Include="ReaderShortcuts.cs" />
+    <Compile Include="UIHandler.cs" />
     <EmbeddedResource Include="Form1.resx">
       <DependentUpon>Form1.cs</DependentUpon>
     </EmbeddedResource>
